feat: extract shared @bragdoc/ai provider package from CLI
Introduce packages/ai (@bragdoc/ai), a shared multi-provider LLM
configuration and factory package consumed by both the CLI and (in
upcoming BYOK work) the web app. It contains the LLMProvider union,
per-provider config interfaces, LLMConfig, DEFAULT_MODELS,
PROVIDER_OPTIONS (names + signup URLs), createLLMFromConfig(),
getLLMDisplayName() and isLLMConfigured(), all moved from the CLI.
The public API is zod-free (CLI is on zod v4, web on zod v3), builds
to dist/ with declarations via tsc, and is published publicly so the
npm-published CLI can depend on it via workspace:*.

The Google provider path no longer mutates
process.env.GOOGLE_GENERATIVE_AI_API_KEY: it now uses
createGoogleGenerativeAI({ apiKey }), matching the explicit-instance
pattern already used for OpenAI, Anthropic and Ollama. This makes the
factory safe for concurrent server-side use while preserving the
config-or-environment-variable key fallback the CLI relies on.

New verifyLLMConfig(config) instantiates the configured model and runs
a tiny generateText probe (maxOutputTokens 8, 15s AbortSignal timeout),
returning { ok: true } or { ok: false, error } with a human-readable
provider error message. It never throws, so the upcoming settings
"verify & save" flow can surface failures directly.

The CLI now imports all of the above from @bragdoc/ai:
createLLMFromConfig/getLLMDisplayName take an LLMConfig (call sites
pass config.llm), config/types.ts re-exports the shared types, and the
duplicated provider/setup code is deleted. lib/extraction.ts no longer
hard-requires OPENAI_API_KEY when a different provider is configured;
missing credentials now fail at model creation with a provider-specific
message.
