+++ .gitignore
@@ -45,6 +45,7 @@ yarn-error.log*
 tmp
 
 .cache_ggshield
+packages/ai/dist
 packages/cli/dist
 
 *.tsbuildinfo
