+++ packages/ai/src/index.ts
@@ -1,0 +1,18 @@
+export {
+  createLLMFromConfig,
+  DEFAULT_MODELS,
+  getLLMDisplayName,
+  isLLMConfigured,
+} from './providers';
+export { PROVIDER_OPTIONS, type ProviderOption } from './provider-options';
+export type {
+  AnthropicConfig,
+  DeepSeekConfig,
+  GoogleConfig,
+  LLMConfig,
+  LLMProvider,
+  OllamaConfig,
+  OpenAICompatibleConfig,
+  OpenAIConfig,
+} from './types';
+export { verifyLLMConfig, type VerifyLLMConfigResult } from './verify';
