+++ packages/cli/package.json
@@ -61,6 +61,7 @@
     "@ai-sdk/anthropic": "^2.0.35",
     "@ai-sdk/google": "^2.0.23",
     "@ai-sdk/openai": "^2.0.53",
+    "@bragdoc/ai": "workspace:*",
     "ai": "^5.0.76",
     "boxen": "^8.0.1",
     "chalk": "^4.1.2",
