+++ packages/cli/src/commands/standup.ts
@@ -16,8 +16,8 @@ import {
 } from '../lib/scheduling';
 import { exec } from 'node:child_process';
 import { promisify } from 'node:util';
-import { promptForLLMConfig, isLLMConfigured } from '../config/llm-setup';
-import { getLLMDisplayName } from '../ai/providers';
+import { getLLMDisplayName, isLLMConfigured } from '@bragdoc/ai';
+import { promptForLLMConfig } from '../config/llm-setup';
 
 const execAsync = promisify(exec);
 
@@ -480,7 +480,7 @@ async function enableStandup() {
       const llmConfig = await promptForLLMConfig();
       config.llm = llmConfig;
       await saveConfig(config);
-      const displayName = getLLMDisplayName(config);
+      const displayName = getLLMDisplayName(config.llm);
       console.log(chalk.green(`✓ LLM configured: ${displayName}\n`));
     }
