+++ packages/cli/src/config/types.ts
@@ -1,3 +1,5 @@
+import type { LLMConfig } from '@bragdoc/ai';
+
 /**
  * Controls what data is extracted from git commits
  */
@@ -68,58 +70,19 @@ export interface StandupProjectConfig {
 }
 
 /**
- * LLM Provider types
+ * LLM Provider types now live in the shared @bragdoc/ai package.
+ * Re-exported here so existing imports keep working.
  */
-export type LLMProvider =
-  | 'openai'
-  | 'anthropic'
-  | 'google'
-  | 'deepseek'
-  | 'ollama'
-  | 'openai-compatible';
-
-export interface OpenAIConfig {
-  apiKey?: string;
-  baseURL?: string;
-  model?: string; // Default: 'gpt-4o'
-}
-
-export interface AnthropicConfig {
-  apiKey?: string;
-  model?: string; // Default: 'claude-3-5-sonnet-20241022'
-}
-
-export interface GoogleConfig {
-  apiKey?: string;
-  model?: string; // Default: 'gemini-1.5-pro'
-}
-
-export interface DeepSeekConfig {
-  apiKey?: string;
-  baseURL?: string;
-  model?: string; // Default: 'deepseek-chat'
-}
-
-export interface OllamaConfig {
-  baseURL?: string; // Default: 'http://localhost:11434'
-  model: string; // Required: e.g., 'llama3.2', 'qwen2.5-coder'
-}
-
-export interface OpenAICompatibleConfig {
-  baseURL: string;
-  apiKey?: string;
-  model: string;
-}
-
-export interface LLMConfig {
-  provider: LLMProvider;
-  openai?: OpenAIConfig;
-  anthropic?: AnthropicConfig;
-  google?: GoogleConfig;
-  deepseek?: DeepSeekConfig;
-  ollama?: OllamaConfig;
-  openaiCompatible?: OpenAICompatibleConfig;
-}
+export type {
+  AnthropicConfig,
+  DeepSeekConfig,
+  GoogleConfig,
+  LLMConfig,
+  LLMProvider,
+  OllamaConfig,
+  OpenAICompatibleConfig,
+  OpenAIConfig,
+} from '@bragdoc/ai';
 
 /**
  * Main configuration type
